removed methods related to expense summary into new ExpenseReportController and added the unit tests into specific ExpenseReportControllerTest
