+++ expense-service/src/main/java/com/example/expense/controller/ExpenseController.java
@@ -48,41 +48,6 @@ public class ExpenseController {
         return ResponseEntity.status(HttpStatus.OK).body(expenseWithRequestedCategory);
 
     }
-
-    @GetMapping("/summary/monthly")
-    public ResponseEntity<ExpenseSummaryResponse> getMonthlyExpenseSummaryResponse(@RequestParam int year,@RequestParam int month){
-        ExpenseSummaryResponse expenseSummaryResponse = new ExpenseSummaryResponse();
-
-        expenseSummaryResponse = expenseService.findExpenseSummaryByMonth(year, month);
-
-        return ResponseEntity.status(HttpStatus.OK).body(expenseSummaryResponse);
-    }
-
-    @GetMapping("/summary/annual")
-    public ResponseEntity<ExpenseSummaryResponse> getAnnualExpenseSummaryResponse(@RequestParam int year){
-        ExpenseSummaryResponse expenseSummaryResponse = new ExpenseSummaryResponse();
-
-        expenseSummaryResponse = expenseService.findExpenseSummaryByYear(year);
-
-        return ResponseEntity.status(HttpStatus.OK).body(expenseSummaryResponse);
-    }
-
-
-    @GetMapping("/detailed/monthly")
-    public ResponseEntity<List<ExpenseResponse>> getMonthlyExpenses(@RequestParam int year,@RequestParam int month){
-        List<ExpenseResponse> expensesFromRequestedMonth = new ArrayList<>();
-
-        expensesFromRequestedMonth = expenseService.findExpensesByMonth(year, month);
-        return ResponseEntity.status(HttpStatus.OK).body(expensesFromRequestedMonth);
-
-    }
-
-    @GetMapping("/detailed/annual")
-    public ResponseEntity<List<ExpenseResponse>> getAnnualExpenses(@RequestParam int year){
-        List<ExpenseResponse> expensesFromRequestedYear = new ArrayList<>();
-            expensesFromRequestedYear = expenseService.findExpensesByYear(year);
-            return ResponseEntity.status(HttpStatus.OK).body(expensesFromRequestedYear);
-    }
     
     @PostMapping
     public ResponseEntity<List<ExpenseResponse>> createExpense(@Valid @RequestBody CreateExpenseRequest expense){
