+++ expense-service/src/test/java/com/example/expense/controller/ExpenseControllerTest.java
@@ -78,99 +78,6 @@ public class ExpenseControllerTest {
 
     }
 
-
-    @Test
-    void getMonthlyExpenseSummaryResponse_ReturnMonthlySummaryResponse_WithMonthlyDetails(){
-        ExpenseSummaryResponse summaryResponse = new ExpenseSummaryResponse();
-
-        summaryResponse.setTotalExpenses(BigDecimal.valueOf(3000));
-
-        summaryResponse.setAverageDailyExpense(BigDecimal.valueOf(100));
-
-        summaryResponse.setHighestSpentCategory("Market");
-
-        List<ExpenseResponse> expenseResponseList = new ArrayList<>();
-        summaryResponse.setDetailedExpenses(expenseResponseList);
-
-        Map<String, BigDecimal> totalPerCategory = new HashMap<>();
-        summaryResponse.setTotalPerCategory(totalPerCategory);
-
-        int year = 2025;
-        int month = 9;
-
-        when(expenseService.findExpenseSummaryByMonth(year, month)).thenReturn(summaryResponse);
-
-        ResponseEntity<ExpenseSummaryResponse> response = expenseController.getMonthlyExpenseSummaryResponse(year, month);
-        assertEquals(HttpStatus.OK, response.getStatusCode());
-        assertEquals(BigDecimal.valueOf(3000), response.getBody().getTotalExpenses());
-        assertEquals(BigDecimal.valueOf(100), response.getBody().getAverageDailyExpense());
-        assertEquals("Market", response.getBody().getHighestSpentCategory());
-        assertEquals(0, response.getBody().getDetailedExpenses().size());
-        assertEquals(0, response.getBody().getTotalPerCategory().size());
-
-
-    }
-
-    @Test
-    void getAnnualExpenseSummaryResponse_ReturnAnnualSummaryResponse_WithAnnualDetails(){
-        ExpenseSummaryResponse summaryResponse = new ExpenseSummaryResponse();
-
-        summaryResponse.setTotalExpenses(BigDecimal.valueOf(3000));
-
-        summaryResponse.setAverageDailyExpense(BigDecimal.valueOf(100));
-
-        summaryResponse.setHighestSpentCategory("Market");
-
-        List<ExpenseResponse> expenseResponseList = new ArrayList<>();
-        summaryResponse.setDetailedExpenses(expenseResponseList);
-
-        Map<String, BigDecimal> totalPerCategory = new HashMap<>();
-        summaryResponse.setTotalPerCategory(totalPerCategory);
-
-        int year = 2025;
-
-        when(expenseService.findExpenseSummaryByYear(year)).thenReturn(summaryResponse);
-
-        ResponseEntity<ExpenseSummaryResponse> response = expenseController.getAnnualExpenseSummaryResponse(year);
-        assertEquals(HttpStatus.OK, response.getStatusCode());
-        assertEquals(BigDecimal.valueOf(3000), response.getBody().getTotalExpenses());
-        assertEquals(BigDecimal.valueOf(100), response.getBody().getAverageDailyExpense());
-        assertEquals("Market", response.getBody().getHighestSpentCategory());
-        assertEquals(0, response.getBody().getDetailedExpenses().size());
-        assertEquals(0, response.getBody().getTotalPerCategory().size());
-
-
-    }
-
-    @Test
-    void getMonthlyExpenses_ReturnExpenseResponseList_WithMonthlyExpenses(){
-        ExpenseResponse expenseResponse = new ExpenseResponse();
-        List<ExpenseResponse> expenseResponseList = new ArrayList<>();
-        expenseResponseList.add(expenseResponse);
-        int year = 2023;
-        int month = 6;
-
-        when(expenseService.findExpensesByMonth(year, month)).thenReturn(expenseResponseList);
-
-        ResponseEntity<List<ExpenseResponse>> response = expenseController.getMonthlyExpenses(year, month);
-        assertEquals(HttpStatus.OK, response.getStatusCode());
-        assertEquals(1, response.getBody().size());
-    }
-
-    @Test
-    void getAnnualExpenses_ReturnExpenseResponseList_WithYearlyExpenses(){
-        ExpenseResponse expenseResponse = new ExpenseResponse();
-        List<ExpenseResponse> expenseResponseList = new ArrayList<>();
-        expenseResponseList.add(expenseResponse);
-        int year = 2025;
-
-        when(expenseService.findExpensesByYear(year)).thenReturn(expenseResponseList);
-
-        ResponseEntity<List<ExpenseResponse>> response = expenseController.getAnnualExpenses(year);
-        assertEquals(HttpStatus.OK, response.getStatusCode());
-        assertEquals(1, response.getBody().size());
-    }
-
     @Test
     void createExpense_ReturnExpenseResponse(){
         CreateExpenseRequest expense = new CreateExpenseRequest();
